pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shadowcheck"
version = "0.1.0"
description = "Autonomous zero-day triage agent with typed output and audit trail"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.8.0",
  "pydantic-ai>=0.0.35",
  "requests>=2.32.0",
  "packaging>=24.0",
  "gradio>=5.0.0",
  "python-dotenv>=1.0.1",
  "logfire>=0.54.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0"
]

[project.scripts]
shadowcheck-cli = "shadowcheck.cli:main"
shadowcheck-ui = "shadowcheck.ui:launch"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
